Don't block on REST call in reaction handler
Previously, `reactionHandler` blocked on the REST call for sending the
message before executing the database call or removing the user from the
newjoin map. This changes it to modify the map first, then fork for the
database call, then block on the rest call as the final action of the
handler.