Add version CLI command
This rewrites the CLI options module to be... better. It adds a
`--version` flag CLI command that prints the version, and a template
Haskell module to provide a splice that reads the version from the cabal
file. Rewrites part of the entry function to use the new interface
exposed by the options module.