Remove TH and clean up entry function
I learned that cabal autogenerates a module with the version number, so
the TH module is unnecessary and has been removed in favor of importing
that. `Main.main` was also cleaned up, including no longer throwing an
exception when there is no Discord token env var and the `--version`
command is invoked.